Refactor a bit more and add more state to the Msg
The password is retreived and then passed on to a couple of Result
combinators. The logic is somewtah like this:

```
read pass -> Validate -ok-> draw(ok) && write pass to file && stop
                      |
                     nok
                      |
                      V
                  draw(fail)
```

The keypress callback is done. It sends a Msg to the drawing thread on
char, enter and escape and only they are handled in the drawing thread.