Make offsets optional, add --image argument
Make x and y offsets optional. The default values of the offsets put the
bgrt image in the center of the screen.

Pass a mutable Framebuffer instance to all drawing functions.

Introduce another cli argument, --image to get the path of the image
displayed on boot. The default value is the bgrt image path. Also, a
Config struct was added to keep the parameters throughout the
application.

Run localy with:
cargo run -- --write pass.txt --image test.bmp